refactor(web): extract local search semantics from hook
Move the local recent filtering, server-vs-local selection, and
request-option shaping from useSessionSearch into pure functions in
lib/search.ts; the hook keeps only UI state and request lifecycle.
Local-path characterization tests migrate to lib-level tests.
